Set a default background color for selects
Typically the browser agent stylesheet sets background-color:white, which is ideal -
but Safari on mobile seems to inherit the background color for selects,
which is not good if it inherits a dark background and also has dark text,
expecting the select to always have a white background. Resolving this
by setting a background-color and not relying on the user agent
stylesheet to do this.